Updated ThreeJS Scene
Added yes and no divs to the scene for options later on. They animate
slide in however there is a problem and they disappear after the
animation is complete. NEED TO FIX.

Otherwise video has been set to autoplay.

Things still to do:
- Make the yes/no options actually look nice and usable.
- Give them the ability to set the VideoTexture to another video source
which is then the arcing story selection part complete.
- Stop the animation from ending and removing the div's.