First Stage Dynamic Page Conversion
I've added in some classes to handle the consumption of JSON and
converted the page to now create the 'Selection' pages based on the JSON
so they are now completely dynamic.
The whole site still works although the rest is still hard coded
(although I removed tones of the random holder stuff we had in there).