Trials for compass data + Fix for Pins + Max zoom added
Some basic code added for getting compass data, being expanded upon as
next priority.
Fixes:
- Changed how the html for the pin content was done so it didn't mess
with any of the styling used elsewhere. The fix was basically adding any
rules in the CSS to only work for children of a unique ID'd parent div.

- Also added a line to stop people from zooming out too far and getting
lost. The level of zoom can be easily changed if its a problem but at
the moment I think its in a nice place. I'll now close Issue #6.